Added a Fixture Import Type.
This new import type allows you to use a schema file from the Config/Schema folder or a plugin Config/Schema folder as your database schema for tests.

This is good because you can have your schema in a single file (as opposed to a schema.php and a fixture file), and still have the benefit of not needing to maintain a separate test database and keep it up to date for importing from a model.